Lockscreen charging info: show decimal in battery temperature

+++ packages/SettingsLib/src/com/android/settingslib/fuelgauge/BatteryStatus.java
@@ -56,12 +56,12 @@ public class BatteryStatus {
     public final int maxChargingWattage;
     public final boolean present;
 
-    public final int temperature;
+    public final float temperature;
 
     public BatteryStatus(int status, int level, int plugged, int health,
             int maxChargingWattage, boolean present,
             int maxChargingCurrent, int maxChargingVoltage,
-            int temperature) {
+            float temperature) {
 
         this.status = status;
         this.level = level;
